refactor: move FilterType.ts
