+++ src/pages/Main/RankingView/FilterType.ts
