+++ src/pages/Main/RankingView/TargetTypeFilter.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 
-import type { TargetType } from '../FilterType';
+import type { TargetType } from './FilterType';
 
 interface Props {
     target: TargetType;
